BLAN-1196 - the list of users in the scheduler is moved to a separate component for reuse
